requirements.txt
flask
requests
beautifulsoup4
google-generativeai>=0.8.3
python-dotenv
playwright

# ========================================
# Cold Outreach CRM Dependencies
# ========================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
sqlalchemy==2.0.25


# Form handling
python-multipart==0.0.6

# Additional utilities
httpx==0.26.0
# Duplicates removed
